mw.ViewPageTarget: Support for non-paragraphs in #showWarning
We need this for captcha's. They'll be added and displayed as
a warning, and cleared like all others. But they shouldn't be
wrapped in a paragraph and aren't delivered as html text but
as a jQuery object (can't serialise to static html since we'll
have to keep a reference to the TextInputWidget and what not).

Change-Id: I3734fbd3c6f0270094db014bbc76d065a73ab918