Avoid self-references, and refer instead to "source mode"
Messages like "This element cannot yet be edited in VisualEditor"
presuppose an awareness on the user's part what VisualEditor is,
but ideally, we want the software to be invisible. Where we want
to refer to behavior of the software, we should characterize it
using language that is discoverable by the user.

Change-Id: Ia4ef7aca1501cf9cf0f772f82ac2b0a99ca53a63